requirements.txt
# Core dependencies for data processing and ML
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Deep learning frameworks
torch>=1.12.0
transformers>=4.20.0
datasets>=2.0.0

# NLP and text processing
nltk>=3.7
spacy>=3.4.0
textblob>=0.17.0

# LLM API clients
openai>=0.27.0
anthropic>=0.3.0

# Jupyter and notebook tools
jupyter>=1.0.0
ipywidgets>=7.7.0
notebook>=6.4.0

# Data visualization
plotly>=5.10.0
wordcloud>=1.8.0

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.28.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
